refactor: modularize test files and eliminate code duplication
- Split form-utils.test.ts (757 lines → 4 focused modules + index)
- Created shared test utilities in test-utils.ts and auth-test-utils.ts
- Eliminated 20 code clones from form-utils tests
- Eliminated 2 code clones from auth.coverage tests
- Maintained 100% test coverage for form-utils.ts
- Fixed ESLint issues with unused parameters

Addresses code quality feedback for file size and duplication.
